requirements.txt
streamlit==1.28.1
python-docx==0.8.11
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.1.3
numpy==1.24.3
jinja2==3.1.2
markdown==3.5.1
PyPDF2==3.0.1 